+++ lib/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
 
-          package="com.nickandjerry.dynamiclayoutinflator.lib"
+          package="org.autojs.dynamiclayoutinflater"
 >
 
     <application android:allowBackup="true"
