--- lib/src/main/java/com/nickandjerry/dynamiclayoutinflator/lib/ViewCreator.java
@@ -1,16 +1,0 @@
-package com.nickandjerry.dynamiclayoutinflator.lib;
-
-import android.content.Context;
-import android.view.View;
-
-import java.util.Map;
-
-/**
- * Created by Stardust on 2017/11/29.
- */
-
-public interface ViewCreator<V extends View> {
-
-    V create(Context context, Map<String, String> attrs);
-
-}
