+++ lib/src/main/java/org/autojs/dynamiclayoutinflater/ViewCreator.java
@@ -1,0 +1,16 @@
+package org.autojs.dynamiclayoutinflater;
+
+import android.content.Context;
+import android.view.View;
+
+import java.util.Map;
+
+/**
+ * Created by Stardust on 2017/11/29.
+ */
+
+public interface ViewCreator<V extends View> {
+
+    V create(Context context, Map<String, String> attrs);
+
+}
