+++ lib/src/main/java/org/autojs/dynamiclayoutinflater/attrsetter/ImageViewAttrSetter.java
@@ -1,0 +1,107 @@
+package org.autojs.dynamiclayoutinflater.attrsetter;
+
+import android.os.Build;
+import android.view.InflateException;
+import android.view.ViewGroup;
+import android.widget.ImageView;
+
+import org.autojs.dynamiclayoutinflater.util.Colors;
+import org.autojs.dynamiclayoutinflater.util.Dimensions;
+import org.autojs.dynamiclayoutinflater.util.Drawables;
+
+import java.util.Map;
+
+/**
+ * Created by Stardust on 2017/11/3.
+ */
+
+public class ImageViewAttrSetter<V extends ImageView> extends BaseViewAttrSetter<V> {
+
+    @Override
+    public boolean setAttr(V view, String attr, String value, ViewGroup parent, Map<String, String> attrs) {
+        if (super.setAttr(view, attr, value, parent, attrs)) {
+            return true;
+        }
+        switch (attr) {
+            case "adjustViewBounds":
+                view.setAdjustViewBounds(Boolean.valueOf(value));
+                break;
+            case "baseline":
+                view.setBaseline(Dimensions.parseToIntPixel(value, view));
+                break;
+            case "baselineAlignBottom":
+                view.setBaselineAlignBottom(Boolean.valueOf(value));
+                break;
+            case "cropToPadding":
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.JELLY_BEAN) {
+                    view.setCropToPadding(Boolean.valueOf(value));
+                }
+                break;
+            case "maxHeight":
+                view.setMaxHeight(Dimensions.parseToIntPixel(value, view));
+                break;
+            case "maxWidth":
+                view.setMaxWidth(Dimensions.parseToIntPixel(value, view));
+                break;
+            case "path":
+                Drawables.setupWithImage(view,  wrapAsPath(value));
+                break;
+            case "scaleType":
+                view.setScaleType(parseScaleType(value));
+                break;
+            case "src":
+                Drawables.setupWithImage(view, value);
+                break;
+            case "tint":
+                view.setColorFilter(Colors.parse(view, value));
+                break;
+            case "tintMode":
+                if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.LOLLIPOP) {
+                    view.setImageTintMode(TINT_MODES.get(value));
+                }
+                break;
+            case "url":
+                Drawables.setupWithImage(view,  wrapAsUrl(value));
+                break;
+            default:
+                return false;
+        }
+        return true;
+    }
+
+    private String wrapAsPath(String value) {
+        if (!value.startsWith("file://")) {
+            return "file://" + value;
+        }
+        return value;
+    }
+
+    private String wrapAsUrl(String value) {
+        if (!value.startsWith("http://") && !value.startsWith("https://")) {
+            return "http://" + value;
+        }
+        return value;
+    }
+
+    private ImageView.ScaleType parseScaleType(String value) {
+        switch (value.toLowerCase()) {
+            case "center":
+                return ImageView.ScaleType.CENTER;
+            case "center_crop":
+                return ImageView.ScaleType.CENTER_CROP;
+            case "center_inside":
+                return ImageView.ScaleType.CENTER_INSIDE;
+            case "fit_center":
+                return ImageView.ScaleType.FIT_CENTER;
+            case "fit_end":
+                return ImageView.ScaleType.FIT_END;
+            case "fit_start":
+                return ImageView.ScaleType.FIT_START;
+            case "fit_xy":
+                return ImageView.ScaleType.FIT_XY;
+            case "matrix":
+                return ImageView.ScaleType.MATRIX;
+        }
+        throw new InflateException("unknown scale type: " + value);
+    }
+}
